bigint 처리를 위한 자료형 변경

+++ Assets/Scripts/GameLogic/GameModel.cs
@@ -75,7 +75,7 @@ public class PlayerModel
         public int level;
         public int awakening;
         public int? guild_id;
-        public int combat_power;
+        public long combat_power;
         public int? rank;
         public int? equipped_skill1_id;
         public int? equipped_skill2_id;
